pyproject.toml
[tool.poetry]
name = "foxwall_api"
version = "0.1.0"
description = ""
authors = ["ümit demir <umitde296@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "=3.2.2"
Pillow = "^8.3.1"
django-imagekit = "^4.0.2"
django-mptt = "^0.12.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
drf-spectacular = "^0.17.3"
PyJWT = "^2.1.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = {version = "^21.7b0", allow-prereleases = true}
autopep8 = "^1.5.7"
yapf = "^0.31.0"
pytest-django = "^4.4.0"
coverage = "^5.5"
pytest-cov = "^2.12.1"
pytest-xdist = "^2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
